api: rename fix/common folder to fix/_common
Rename the module to indicate that the content in this folder
is for internal use only
