+++ uaclient/api/tests/test_api_u_pro_security_fix_execute.py
@@ -22,7 +22,7 @@ from uaclient.api.u.pro.security.fix import (
     PackageCannotBeInstalledData,
     SecurityIssueNotFixedData,
 )
-from uaclient.api.u.pro.security.fix.common.execute.v1 import (
+from uaclient.api.u.pro.security.fix._common.execute.v1 import (
     FailedUpgrade,
     FixExecuteError,
     FixExecuteResult,
