+++ uaclient/api/u/pro/security/fix/cve/plan/v1.py
@@ -3,7 +3,7 @@ from typing import List
 from uaclient.api.api import APIEndpoint
 from uaclient.api.data_types import AdditionalInfo
 from uaclient.api.u.pro.security.fix import FixPlanResult, fix_plan_cve
-from uaclient.api.u.pro.security.fix.common import get_expected_overall_status
+from uaclient.api.u.pro.security.fix._common import get_expected_overall_status
 from uaclient.config import UAConfig
 from uaclient.data_types import DataObject, Field, StringDataValue, data_list
