+++ uaclient/cli/fix.py
@@ -42,7 +42,7 @@ from uaclient.api.u.pro.security.fix import (  # noqa: F401
     NoOpLivepatchFixData,
     USNAdditionalData,
 )
-from uaclient.api.u.pro.security.fix.common import status_message
+from uaclient.api.u.pro.security.fix._common import status_message
 from uaclient.api.u.pro.security.fix.cve.plan.v1 import CVEFixPlanOptions
 from uaclient.api.u.pro.security.fix.cve.plan.v1 import _plan as cve_plan
 from uaclient.api.u.pro.security.fix.usn.plan.v1 import USNFixPlanOptions
